requirements.txt
# Core App
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# PDF/DOCX Parsing
pdfminer.six==20221105
python-docx==1.1.0

# Embeddings & LLMs
sentence-transformers==2.7.0
InstructorEmbedding==1.0.1
faiss-cpu==1.7.4
google-generativeai==0.8.5

# Utils
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.9.5

# Logging
loguru==0.7.2

# Testing
pytest==8.2.2
httpx==0.27.0

# Streamlit UI
streamlit==1.35.0
PyMuPDF

